Add data augmentation and grayscale support
Major features added. Support grayscale version on GENKI-4K. Add
dataaugmentation for both color and grayscale versions. Add option to
visualize data. Add option for training. Add option for serialization.
The goal of this commit was to enable all the functionality via the
main.py file. There is no longer a need to use the dataloader.py for serialization.